+++ packages/core/src/components/Combobox/__tests__/Combobox.test.js
@@ -21,7 +21,8 @@ describe("Combobox tests", () => {
   describe("Without categories", () => {
     const mockOptions = [
       { value: "orange", label: "Orange" },
-      { value: "yellow", label: "Yellow" }
+      { value: "yellow", label: "Yellow" },
+      { value: "red", label: "Red", disabled: true }
     ];
 
     it("should call item on click callback func when onClick", () => {
@@ -88,6 +89,22 @@ describe("Combobox tests", () => {
         fireEvent.click(screen.getByText("Add new"));
         expect(onAddMock.mock.calls.length).toBe(1);
       });
+    });
+
+    it("should not call onClick for disabled option", () => {
+      const onClickMock = jest.fn();
+      const { getByLabelText } = render(<Combobox onClick={onClickMock} options={mockOptions} />);
+
+      fireEvent.click(getByLabelText("Red"));
+      expect(onClickMock).not.toHaveBeenCalled();
+    });
+
+    it("should NOT trigger hover event on disabled options", () => {
+      const onHoverMock = jest.fn();
+      const { getByLabelText } = render(<Combobox onOptionHover={onHoverMock} options={mockOptions} />);
+
+      fireEvent.mouseOver(getByLabelText("Red"));
+      expect(onHoverMock).not.toHaveBeenCalled();
     });
   });