+++ packages/core/src/components/Dropdown/Dropdown.styles.js
@@ -39,7 +39,11 @@ const getColor = () => {
   return { color, backgroundColor };
 };
 
-const getFont = size => ({ fontSize: getSingleValueTextSize(size), lineHeight: getSingleValueTextSize(size) });
+const getFont = size => ({
+  fontSize: getSingleValueTextSize(size),
+  lineHeight: getSingleValueTextSize(size),
+  fontFamily: getCSSVar("font-family")
+});
 
 const disabledContainerStyle = isDisabled => {
   if (!isDisabled) return {};
@@ -320,7 +324,9 @@ const menu =
       ...getFont(),
       color: getCSSVar("primary-text-color"),
       backgroundColor: getCSSVar("dialog-background-color"),
-      boxShadow: getCSSVar("experimental-dropdown-border-in-dark-themes", getCSSVar("box-shadow-small"))
+      boxShadow: `${getCSSVar("experimental-dropdown-border-in-dark-themes", "0 0 0 0 transparent")}, ${getCSSVar(
+        "box-shadow-small"
+      )}`
     };
 
     if (!insideOverflowContainer && !insideOverflowWithTransformContainer) return baseStyle;