+++ packages/core/src/components/Dropdown/Dropdown.tsx
@@ -2,23 +2,24 @@
 import { ComponentDefaultTestId, getTestId } from "../../tests/test-ids-utils";
 import cx from "classnames";
 import { SIZES, SIZES_VALUES } from "../../constants";
-import React, { forwardRef, useCallback, useMemo, useRef, useState } from "react";
+import React, { forwardRef, useCallback, useMemo, useRef, useState, useEffect } from "react";
 import Select, { InputProps, components, createFilter, ActionMeta } from "react-select";
 import AsyncSelect from "react-select/async";
 import BaseSelect from "react-select/base";
 import { noop as NOOP } from "lodash-es";
-import { WindowedMenuList } from "react-windowed-select";
 import MenuComponent from "./components/menu/menu";
 import DropdownIndicatorComponent from "./components/DropdownIndicator/DropdownIndicator";
 import OptionComponent from "./components/option/option";
 import SingleValueComponent from "./components/singleValue/singleValue";
 import ClearIndicatorComponent from "./components/ClearIndicator/ClearIndicator";
 import MultiValueContainer from "./components/MultiValueContainer/MultiValueContainer";
+import { isClient } from "../../utils/ssr-utils";
 import {
   ADD_AUTO_HEIGHT_COMPONENTS,
   defaultCustomStyles,
   DROPDOWN_CHIP_COLORS,
   DROPDOWN_ID,
+  DROPDOWN_INPUT_ARIA_LABEL,
   DROPDOWN_MENU_ARIA_LABEL,
   DROPDOWN_MENU_ID,
   DROPDOWN_MENU_PLACEMENT,
@@ -60,8 +61,11 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
       onMenuClose = NOOP,
       onFocus = NOOP,
       onBlur = NOOP,
+      onScroll = NOOP,
+      onMenuScrollToBottom = NOOP,
       onChange: customOnChange = NOOP,
       searchable = true,
+      captureMenuScroll = false,
       options = [],
       defaultValue,
       value: customValue,
@@ -89,6 +93,7 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
       id = DROPDOWN_ID,
       menuId = DROPDOWN_MENU_ID,
       menuAriaLabel = DROPDOWN_MENU_ARIA_LABEL,
+      inputAriaLabel = DROPDOWN_INPUT_ARIA_LABEL,
       autoFocus = false,
       multi = false,
       multiline = false,
@@ -132,6 +137,17 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
     BaseSelect.prototype.renderLiveRegion = () => {
       return null;
     };
+
+    // SSR support
+    const [WindowedMenuList, setWindowedMenuList] = useState(null);
+    useEffect(() => {
+      if (isClient()) {
+        // Dynamically import the specific named export from react-windowed-select for SSR support
+        import("react-windowed-select").then(module => {
+          setWindowedMenuList(() => module.WindowedMenuList);
+        });
+      }
+    }, []);
 
     const [selected, setSelected] = useState(overrideDefaultValue || []);
     const [focusedOptionId, setFocusedOptionId] = useState("");
@@ -214,9 +230,10 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
           ariaLabel={menuAriaLabel}
           Renderer={menuRenderer}
           dropdownMenuWrapperClassName={dropdownMenuWrapperClassName}
+          onScroll={onScroll}
         />
       ),
-      [dropdownMenuWrapperClassName, menuRenderer, menuId, menuAriaLabel]
+      [dropdownMenuWrapperClassName, menuRenderer, menuId, menuAriaLabel, onScroll]
     );
 
     const DropdownIndicator = useCallback(
@@ -248,7 +265,7 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
             aria-activedescendant={ariaActiveDescendant}
             role="combobox"
             aria-expanded={!readOnly && menuIsOpen}
-            aria-label="Dropdown input"
+            aria-label={inputAriaLabel}
             aria-controls={menuId}
           />
         );
@@ -391,7 +408,7 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
             MultiValue: NOOP, // We need it for react-select to behave nice with "multi"
             ValueContainer: MultiValueContainer
           }),
-          ...(isVirtualized && { MenuList: WindowedMenuList })
+          ...(isVirtualized && WindowedMenuList && { MenuList: WindowedMenuList })
         }}
         // When inside scroll we set the menu position by js and we can't follow the drop down location while use scrolling
         closeMenuOnScroll={closeMenuOnScroll}
@@ -413,6 +430,8 @@ const Dropdown: VibeComponent<DropdownComponentProps, HTMLElement> & {
         onMenuClose={onMenuClose}
         onFocus={onFocus}
         onBlur={onBlur}
+        onMenuScrollToBottom={onMenuScrollToBottom}
+        captureMenuScroll={captureMenuScroll}
         onChange={onChange}
         onKeyDown={onKeyDown}
         onInputChange={onInputChange}