+++ packages/core/src/components/Dropdown/DropdownConstants.ts
@@ -5,6 +5,7 @@ export const ADD_AUTO_HEIGHT_COMPONENTS = ["container", "control", "valueContain
 export const DROPDOWN_ID = "dropdown-menu-id";
 export const DROPDOWN_MENU_ID = "dropdown-menu-list-id";
 export const DROPDOWN_MENU_ARIA_LABEL = "Dropdown menu";
+export const DROPDOWN_INPUT_ARIA_LABEL = "Dropdown input";
 
 export const DROPDOWN_CHIP_COLORS = {
   PRIMARY: "PRIMARY",