+++ packages/core/src/components/Icon/Icon.tsx
@@ -38,7 +38,7 @@ export interface IconProps extends VibeComponentProps {
    */
   clickable?: boolean;
   /**
-mo   * Icon aria label [aria label](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-label)
+   * Icon aria label [aria label](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-label)
    */
   iconLabel?: string;
   /**