+++ packages/core/src/components/Table/context/TableContext/TableContext.types.ts
@@ -1,24 +1,30 @@
 import { ITableProps } from "../../Table/Table";
 import React, { UIEventHandler } from "react";
 
-export interface ITableContext {
+export interface ITableContext extends TableProviderValue {
+  headRef: React.MutableRefObject<HTMLDivElement>;
+  onHeadScroll: UIEventHandler<HTMLDivElement>;
+  virtualizedListRef: React.MutableRefObject<HTMLDivElement>;
+  onVirtualizedListScroll: UIEventHandler<HTMLDivElement>;
+}
+
+export interface TableProviderForwardedProps {
   columns: ITableProps["columns"];
   dataState?: ITableProps["dataState"];
   emptyState: ITableProps["emptyState"];
   errorState: ITableProps["errorState"];
   size: ITableProps["size"];
+}
+
+export interface TableProviderValue extends TableProviderForwardedProps {
   tableRootRef: React.MutableRefObject<HTMLDivElement>;
-  scrollLeft: number;
-  onTableRootScroll: UIEventHandler<HTMLDivElement>;
-  headRef: React.MutableRefObject<HTMLDivElement>;
-  onHeadScroll: UIEventHandler<HTMLDivElement>;
-  virtualizedListRef: React.MutableRefObject<HTMLDivElement>;
-  onVirtualizedListScroll: UIEventHandler<HTMLDivElement>;
   isVirtualized: boolean;
   markTableAsVirtualized: () => void;
+  scrollLeft: number;
+  setScrollLeft: (scrollAmount: number) => void;
 }
 
 export type ITableProviderProps = {
-  value: Pick<ITableContext, "columns" | "dataState" | "emptyState" | "errorState" | "size">;
+  value: TableProviderValue;
   children: React.ReactNode;
 };