+++ packages/style/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.16.0](https://github.com/mondaycom/vibe/compare/monday-ui-style@0.15.0...monday-ui-style@0.16.0) (2024-10-31)
+
+
+### Features
+
+* **Clipboard:** new icon ([#2570](https://github.com/mondaycom/vibe/issues/2570)) ([da71c1b](https://github.com/mondaycom/vibe/commit/da71c1bc6ca23a5321b2ea6e177633da6d2e01cf))
+* **Forward:** new icon ([#2569](https://github.com/mondaycom/vibe/issues/2569)) ([2034fd5](https://github.com/mondaycom/vibe/commit/2034fd51e30e2c32425e0e06322739d3526c22eb))
+
+
+
+
+
+# [0.15.0](https://github.com/mondaycom/vibe/compare/monday-ui-style@0.14.1...monday-ui-style@0.15.0) (2024-10-29)
+
+
+### Features
+
+* **Wand:** add "Template" tag to icon ([#2565](https://github.com/mondaycom/vibe/issues/2565)) ([9909743](https://github.com/mondaycom/vibe/commit/99097433d742bd77a94b72968b3ed2008dbf769e))
+
+
+
+
+
+## [0.14.1](https://github.com/mondaycom/vibe/compare/monday-ui-style@0.14.0...monday-ui-style@0.14.1) (2024-09-29)
+
+**Note:** Version bump only for package monday-ui-style
+
+
+
+
+
 # [0.14.0](https://github.com/mondaycom/vibe/compare/monday-ui-style@0.13.0...monday-ui-style@0.14.0) (2024-09-26)